# cmd/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/admin//product/orderstatus": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "get order status list for admin",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin.OrderDash"
                ],
                "summary": "API for get order status list",
                "operationId": "GetOrderStatus",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.OrderStatus"
                            }
                        }
                    },
                    "500": {
                        "description": "Something went wrong !"
                    }
                }
            }
        },
...
        "/admin/product/EditProduct": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "Admin.ProductDash"
                ],
                "summary": "api for admin to edit product",
                "operationId": "EditProduct",
                "parameters": [
                    {
                        "description": "inputs",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.EditProduct"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "edited successfully"
                    },
                    "400": {
                        "description": "Missing or invalid entry"
                    }
                }
            }
        },
        "/admin/product/addcategory": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "Admin.ProductDash"
                ],
                "summary": "api for admin to add catogory",
                "operationId": "AddCategory",
                "parameters": [
                    {
                        "description": "inputs",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.AddCatogory"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "category added successfully"
                    },
                    "400": {
                        "description": "Missing or invalid entry"
                    },
                    "500": {
                        "description": "Something went wrong !"
                    }
                }
            }
        },
        "/admin/product/addorderstatus": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "Admin.OrderDash"
                ],
                "summary": "api for admin to add status",
                "operationId": "AddOrderStatus",
                "parameters": [
                    {
                        "description": "inputs",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.OrderStatus"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "status added successfully"
                    },
                    "400": {
                        "description": "Missing or invalid entry"
                    },
                    "500": {
                        "description": "Something went wrong !"
                    }
                }
            }
        },
        "/admin/product/addprice": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "Admin.ProductDash"
                ],
                "summary": "api for admin to add price",
                "operationId": "AddPrice",
                "parameters": [
                    {
                        "description": "inputs",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.AddPrice"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "price added successfully"
                    },
                    "400": {
                        "description": "Missing or invalid entry"
                    }
                }
            }
        },
        "/admin/product/addproduct": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "Admin.ProductDash"
                ],
                "summary": "api for admin to add product",
                "operationId": "AddProduct",
                "parameters": [
                    {
                        "description": "inputs",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.AddProduct"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "product added successfully"
                    },
                    "400": {
                        "description": "Missing or invalid entry"
                    },
                    "500": {
                        "description": "Something went wrong !"
                    }
                }
            }
        },
        "/admin/product/addvariation": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "Admin.ProductDash"
                ],
                "summary": "api for admin to add varient",
                "operationId": "AddVarient",
                "parameters": [
                    {
                        "description": "inputs",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.AddVarient"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "varient added successfully"
                    },
                    "400": {
                        "description": "Missing or invalid entry"
                    },
                    "500": {
                        "description": "Something went wrong !"
                    }
                }
            }
        },
        "/admin/product/editprice": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "Admin.ProductDash"
                ],
                "summary": "api for admin to edit price",
                "operationId": "EditPrice",
                "parameters": [
                    {
                        "description": "inputs",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.EditPrice"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "edited successfully"
                    },
                    "400": {
                        "description": "Missing or invalid entry"
                    },
                    "500": {
                        "description": "Something went wrong !"
                    }
                }
            }
        },
        "/admin/product/variation": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "get variation list for admin and user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin.ProductDash"
                ],
                "summary": "API for get variation list",
                "operationId": "Getvariations",
                "responses": {
                    "200": {
                        "description": "Successfully grt variation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Variation"
                            }
                        }
                    },
                    "500": {
                        "description": "Something went wrong !"
                    }
                }
            }
        },
...
        "/user/product/": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User.Product",
                    "Admin.ProductDash"
                ],
                "summary": "API for admin or user to list all products",
                "operationId": "ProductList",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page Number",
                        "name": "page_number",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Count Of Order",
                        "name": "count",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ProductRes"
                        }
                    },
                    "204": {
                        "description": "No products to show"
                    },
                    "500": {
                        "description": "Failed to get all products"
                    }
                }
            }
        },
        "/user/product/category": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "get category list for admin and user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin.ProductDash",
                    "User.Product"
                ],
                "summary": "API for get category list",
                "operationId": "GetCategory",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Category"
                            }
                        }
                    },
                    "400": {
                        "description": "string \"Invalid input"
                    }
                }
            }
        },
        "/user/product/listproductbyid": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "GetProduct list for admin and user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin.ProductDash",
                    "User.Product"
                ],
                "summary": "Get a product by ID",
                "operationId": "GetProduct",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "productID",
                        "name": "ID",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ProductDetails"
                        }
                    },
                    "400": {
                        "description": "can't get product",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/product/listproductsbycatogory": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "GetProduct list for admin and user by using name",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin.ProductDash",
                    "User.Product"
                ],
                "summary": "Get a product by name",
                "operationId": "GetProductsByCategoryName",
                "parameters": [
                    {
                        "type": "string",
                        "description": "category",
                        "name": "Name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ProductDetails"
                        }
                    },
                    "204": {
                        "description": "didnt get catogory name"
                    },
                    "400": {
                        "description": "can't get product",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.Category": {
            "type": "object",
            "properties": {
                "category_name": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "domain.OrderStatus": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 3
                }
            }
        },
...
        "domain.Variation": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "description": "VariationIndex uint      `json:\"variation_index\" gorm:\"unique;not null\"`",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "request.AddCatogory": {
            "type": "object",
            "required": [
                "category_name"
            ],
            "properties": {
                "category_name": {
                    "type": "string",
                    "maxLength": 15,
                    "minLength": 3
                }
            }
        },
        "request.AddPrice": {
            "type": "object",
            "required": [
                "discount_price",
                "price",
                "product_ID",
                "variation_id"
            ],
            "properties": {
                "discount_price": {
                    "type": "number"
                },
                "price": {
                    "type": "number"
                },
                "product_ID": {
                    "type": "integer"
                },
                "variation_id": {
                    "type": "integer"
                }
            }
        },
        "request.AddProduct": {
            "type": "object",
            "required": [
                "Code",
                "category_id",
                "description",
                "image",
                "product_name",
                "qty_in_stock"
            ],
            "properties": {
                "Code": {
                    "type": "integer"
                },
                "category_id": {
                    "type": "integer"
                },
                "description": {
                    "type": "string",
                    "maxLength": 400,
                    "minLength": 2
                },
                "image": {
                    "type": "string"
                },
                "product_name": {
                    "type": "string",
                    "maxLength": 15,
                    "minLength": 3
                },
                "qty_in_stock": {
                    "type": "integer"
                }
            }
        },
        "request.AddVarient": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "maxLength": 15,
                    "minLength": 3
                }
            }
        },
...
        "request.EditPrice": {
            "type": "object",
            "required": [
                "discount_price",
                "price",
                "product_ID",
                "variation_id"
            ],
            "properties": {
                "discount_price": {
                    "type": "number"
                },
                "price": {
                    "type": "number"
                },
                "product_ID": {
                    "type": "integer"
                },
                "variation_id": {
                    "type": "integer"
                }
            }
        },
        "request.EditProduct": {
            "type": "object",
            "required": [
                "Code"
            ],
            "properties": {
                "Code": {
                    "type": "integer"
                },
                "category_id": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "product_name": {
                    "type": "string"
                },
                "qty_in_stock": {
                    "type": "integer"
                },
                "stock_status": {
                    "type": "boolean"
                }
            }
        },
...
        "response.ProductRes": {
            "type": "object",
            "properties": {
                "category_name": {
                    "type": "string"
                },
                "code": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image": {
                    "type": "string"
                },
                "product_name": {
                    "type": "string"
                },
                "qty_in_stock": {
                    "type": "integer"
                }
            }
        },
...
    }
}